refactor: extract error guidance constants into separate module
Extract EXIT_CODE_GUIDANCE, SIGNAL_GUIDANCE, and NETWORK_ERROR_GUIDANCE
data structures from commands.ts into a new error-guidance.ts module. This
reduces commands.ts by ~260 lines and improves separation of concerns.

Key changes:
- Created new cli/src/error-guidance.ts module for all error guidance constants
- Moved interfaces (SignalEntry, ExitCodeEntry, ErrorGuidance) to error-guidance.ts
- Moved helper functions (buildDashboardHint, optionalDashboardLine) to error-guidance.ts
- Renamed exported functions for clarity (getSignalGuidance → getSignalGuidanceLines, getScriptFailureGuidance → getExitCodeGuidanceLines)
- Added backward-compatible re-exports in commands.ts for any external consumers
- Resolved circular dependency by using setCredentialHintsFn callback pattern

Commands.ts now focuses on CLI dispatch and flow control, while error-guidance.ts
owns all structured error messaging logic.
